requirements.txt
boltz==2.0.3
chembl-structure-pipeline==1.2.2
chembl-webresource-client==0.10.9
faiss-cpu==1.10.0
faiss-gpu==1.7.2
git+https://github.com/MolecularAI/REINVENT4.git@REINVENT4#egg=reinvent4
langchain==0.3.23
langchain-anthropic==0.3.13
langchain-community==0.3.21
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.25
rdkit==2024.9.6
requests==2.32.3
torch==2.6.0
transformers==4.41.0